Visual Basic Units Conversion Application
Simple application written in VB.NET allowing 8 different conversion
options between common units. This includes ounces to grams, litres to
gallons, kilograms to stones, celsius to fahrenheit and vice versa. Also
allows a table to be produced to show a range of values being converted.
This application has been written with the aim to shorten the length of
the code as much as possible.